+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Breaking changes in **bold**.
   - eliminate `mountEvents` stream `MountEvent`, and all related machinery
   - New: onMount / onUnmount / onLifecycle methods
   - Order of unmount events might have changed (see test)
+* **API: eliminate auxiliary syntax `myElement <-- child <-- childSignal`**
+  * Use the new `amend` method instead: `myElement.amend(child <-- childSignal)`
 * **API: Lifecycle events & Ownership overhaul**
   * <TODO: Elaborate> Use Airstream's new `DynamicOwner` and `DynamicSubscription` features: `ReactiveElement` subscriptions are now activated only when the element is mounted, not when it is created.
   * <TODO: Elaborate> Transaction delay for `mountEvents` now applies to all element subscriptions